Fix out-of-bounds memory access in TensorFlow Lite dimensions scaling logic by upgrading dimension calculation index bindings to 64-bit integer types.
PiperOrigin-RevId: 926144762

+++ tensorflow/lite/kernels/internal/optimized/legacy_optimized_ops.h
@@ -4998,8 +4998,9 @@ void ArgMax(const T3* axis, const T1* input_data,
             const tflite::Dims<4>& output_dims) {
   // Assumes the input always has 4 dimensions, and therefore,
   // output always has three dimensions.
-  auto output_shape = RuntimeShape(
-      {output_dims.sizes[2], output_dims.sizes[1], output_dims.sizes[0]});
+  auto output_shape = RuntimeShape({static_cast<int>(output_dims.sizes[2]),
+                                    static_cast<int>(output_dims.sizes[1]),
+                                    static_cast<int>(output_dims.sizes[0])});
   // Another way to interpret this is that output_dims.sizes[4] is always 1.
   TFLITE_DCHECK_EQ(output_shape.FlatSize(),
                    DimsToShape(output_dims).FlatSize());
